R6 Class for Gradient Boosting

An R6 Class that provides an interface to gradient boosting with neural network feature transformation.

Public fields

estimators

List of fitted base models

learning_rate

Learning rate for boosting

losses

Vector of training losses

n_estimators

Number of estimators used

Methods

Public methods

Booster$new()

Booster$fit()

Booster$predict()

Booster$clone()

Method new()

Create a new Booster object

Usage

Booster$new(
  model_name = "ExtraTreeRegressor",
  n_estimators = 100L,
  learning_rate = 0.1,
  tolerance = 1e-04,
  calibration = FALSE,
  seed = 123L,
  show_progress = TRUE,
  verbose = FALSE,
  venv_path = "./venv"
)

Arguments

model_name

Name of the base model

n_estimators

Number of boosting iterations

learning_rate

Learning rate for boosting

tolerance

Convergence tolerance

calibration

Whether to calibrate the model

seed

Random seed

show_progress

Whether to show progress bar

verbose

Whether to print detailed output

venv_path

Path to the virtual environment

Method fit()

Fit the boosting model to training data

Usage

Booster$fit(x, y)

Arguments

x

Feature matrix

y

Target vector

Returns

The fitted object (invisible)

Method predict()

Make predictions on new data

Usage

Booster$predict(newdata)

Arguments

newdata

New data to predict on

Returns

Vector of predictions

Method clone()

The objects of this class are cloneable with this method.

Usage

Booster$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.