Get a list of all models in scikit-learn

This function retrieves a list of all models available in scikit-learn. It imports the necessary Python modules and retrieves all estimators, filtering them into classifiers and regressors.

get_model_list()

Value

A list with two elements: - `classifiers`: A character vector of all classifier models - `regressors`: A character vector of all regressor models

Examples

model_list <- get_model_list()
print(model_list$classifiers)
#>  [1] "AdaBoostClassifier"             "BaggingClassifier"             
#>  [3] "BernoulliNB"                    "CalibratedClassifierCV"        
#>  [5] "CategoricalNB"                  "ClassifierChain"               
#>  [7] "ComplementNB"                   "DecisionTreeClassifier"        
#>  [9] "DummyClassifier"                "ExtraTreeClassifier"           
#> [11] "ExtraTreesClassifier"           "FixedThresholdClassifier"      
#> [13] "GaussianNB"                     "GaussianProcessClassifier"     
#> [15] "GradientBoostingClassifier"     "HistGradientBoostingClassifier"
#> [17] "KNeighborsClassifier"           "LabelPropagation"              
#> [19] "LabelSpreading"                 "LinearDiscriminantAnalysis"    
#> [21] "LinearSVC"                      "LogisticRegression"            
#> [23] "LogisticRegressionCV"           "MLPClassifier"                 
#> [25] "MultiOutputClassifier"          "MultinomialNB"                 
#> [27] "NearestCentroid"                "NuSVC"                         
#> [29] "OneVsOneClassifier"             "OneVsRestClassifier"           
#> [31] "OutputCodeClassifier"           "PassiveAggressiveClassifier"   
#> [33] "Perceptron"                     "QuadraticDiscriminantAnalysis" 
#> [35] "RadiusNeighborsClassifier"      "RandomForestClassifier"        
#> [37] "RidgeClassifier"                "RidgeClassifierCV"             
#> [39] "SGDClassifier"                  "SVC"                           
#> [41] "SelfTrainingClassifier"         "StackingClassifier"            
#> [43] "TunedThresholdClassifierCV"     "VotingClassifier"